Fix multiple moves, turns; rework monster_group
sprite.group.copy(), probably because of python's wierd pass by referene rules,
or possibly scoping rules, was not working as a copy, but rather as a pointer to
the same object, so we were removing everything from the monster_group by
accident. The other thing is that we really want to use just one monster_group
because when we kill a sprite, we want all other monsters to be aware. So, now
we use one group and then add and remove a monster's self from the group when
checking for monster-monster collisions.

Fix multiple moves.

Monsters were not looking up the damage from M_DATA[][] when initializing.